gs1(serializer)/requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.1
django-filter==2.2.0
django-suit==2.0a1
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
